updated avatar hooks
