+++ src/pages/settings/avatar/Avatar.js
@@ -2,9 +2,11 @@ import styles from "./Avatar.module.css";
 
 import { useState } from "react";
 
-import { useUploadAvatar } from "../../../hooks/avatar/useUploadAvatar";
 import { useReadProfile } from "../../../hooks/user/userApis";
-import { useDeleteAvatar } from "../../../hooks/avatar/useDeleteAvatar";
+import {
+  useDeleteAvatar,
+  useUploadAvatar,
+} from "../../../hooks/avatar/avatarApis";
 
 import { useSelector } from "react-redux";
